requirements.txt
snowflake.sqlalchemy
pandas
dotenv
streamlit 
snowflake-snowpark-python
plotly 
